refactor(mfa): rename reachableStates to settleableLeafStates
The module enumerates settleable leaf states of the chart and computes
no reachability (the specs prove that). The new name matches its only
export, getSettleableLeafStates.
